Extract the bulk search code into a class
This isn't pretty yet, but it makes refactoring easier.
